Enforce strong typing for API parameters and data models

+++ Telegram.BotAPI/Parameters/UpdatingMessages/ApproveSuggestedPostParameters.cs
@@ -4,7 +4,7 @@ public sealed class ApproveSuggestedPostParameters : ApiRequestParameters
 {
     public required long ChatId { get; init; }
 
-    public required int MessageId { get; init; }
+    public required long MessageId { get; init; }
 
     public int? SendDate { get; init; }
 }
