+++ Telegram.BotAPI/Parameters/UpdatingMessages/DeclineSuggestedPostParameters.cs
@@ -4,7 +4,7 @@ public sealed class DeclineSuggestedPostParameters : ApiRequestParameters
 {
     public required long ChatId { get; init; }
 
-    public required int MessageId { get; init; }
+    public required long MessageId { get; init; }
 
     public string? Comment { get; init; }
 }
